Move tests and fix linting
