+++ tests/integration/common.test.js
@@ -1,36 +1,32 @@
-import { toggleAttribute, toggleClass } from './common';
-
-// DOM Elements to be set
-let testElement;
-const undefinedElement = document.querySelector('.nhs-false-element');
-
-// Helper to set DOM Elements
-const initTest = (html = '<div class="nhsuk-test" />') => {
-  document.body.innerHTML = html;
-  testElement = document.querySelector('.nhsuk-test');
-};
+import { toggleAttribute, toggleClass } from '../../packages/common';
 
 describe('toggleAttribute util', () => {
+  const attr = 'aria-expanded';
+  let testElement;
+
+  beforeEach(() => {
+    document.body.innerHTML = '<div class="nhsuk-test" />';
+    testElement = document.querySelector('.nhsuk-test');
+  });
+
   describe('Does not throw an error', () => {
     it('if the element does not exist', () => {
-      initTest('');
+      const undefinedElement = document.querySelector('.nhs-false-element');
       expect(undefinedElement).toBeNull();
-      toggleAttribute(document.querySelector('.nhs-false-element'), 'aria-expanded');
+      toggleAttribute(undefinedElement, 'aria-expanded');
     });
   });
 
   describe('Adds attribute with value of "true"', () => {
-    const attr = 'aria-expanded';
 
     it('if the element currently has no attribute', () => {
-      initTest();
       expect(testElement).toBeDefined();
       toggleAttribute(testElement, attr);
       expect(testElement.getAttribute(attr)).toEqual('true');
     });
 
     it('if the attributes current value is "false"', () => {
-      initTest(`<div class="nhsuk-test" ${attr}="false" />`);
+      document.body.innerHTML = `<div class="nhsuk-test" ${attr}="false" />`;
       expect(testElement).toBeDefined();
       toggleAttribute(testElement, attr);
       expect(testElement.getAttribute(attr)).toEqual('true');
@@ -38,10 +34,9 @@ describe('toggleAttribute util', () => {
   });
 
   describe('Adds attribute with value of "false"', () => {
-    const attr = 'aria-expanded';
-
     it('if the attributes current value is "true"', () => {
-      initTest(`<div class="nhsuk-test" ${attr}="true" />`);
+      document.body.innerHTML = `<div class="nhsuk-test" ${attr}="true" />`;
+      testElement = document.querySelector('.nhsuk-test');
       expect(testElement).toBeDefined();
       toggleAttribute(testElement, attr);
       expect(testElement.getAttribute(attr)).toEqual('false');
@@ -51,10 +46,16 @@ describe('toggleAttribute util', () => {
 
 describe('toggleClass util', () => {
   const className = 'test-class';
+  let testElement;
+
+  beforeEach(() => {
+    document.body.innerHTML = '<div class="nhsuk-test" />';
+    testElement = document.querySelector('.nhsuk-test');
+  });
 
   describe('Does not throw an error', () => {
     it('if the element does not exist', () => {
-      initTest('');
+      const undefinedElement = document.querySelector('.nhs-false-element');
       expect(undefinedElement).toBeNull();
       toggleClass(undefinedElement, 'test-class');
     });
@@ -62,7 +63,6 @@ describe('toggleClass util', () => {
 
   describe('Adds class', () => {
     it('if the class does not already exist on the element', () => {
-      initTest();
       expect(testElement).toBeDefined();
       toggleClass(testElement, className);
       expect(testElement.classList.contains(className)).toEqual(true);
@@ -71,7 +71,8 @@ describe('toggleClass util', () => {
 
   describe('Removes class', () => {
     it('if the class already exists on the element', () => {
-      initTest(`<div class="nhsuk-test ${className}" />`);
+      document.body.innerHTML = `<div class="nhsuk-test ${className}" />`;
+      testElement = document.querySelector('.nhsuk-test');
       expect(testElement).toBeDefined();
       toggleClass(testElement, className);
       expect(testElement.classList.contains(className)).toEqual(false);
